+++ src/include/mk_http.h
@@ -79,11 +79,11 @@ int mk_http_method_get(char *body);
 int mk_http_protocol_check(char *protocol, int len);
 mk_ptr_t mk_http_protocol_check_str(int protocol);
 
-int mk_http_init(struct client_session *cs, struct session_request *sr);
+ssize_t mk_http_init(struct client_session *cs, struct session_request *sr);
 int mk_http_keepalive_check(struct client_session *cs);
 
 int mk_http_pending_request(struct client_session *cs);
-int mk_http_send_file(struct client_session *cs, struct session_request *sr);
+ssize_t mk_http_send_file(struct client_session *cs, struct session_request *sr);
 int mk_http_request_end(int socket);
 
 #endif
