+++ src/include/mk_socket.h
@@ -68,8 +68,8 @@ int mk_socket_accept(int server_fd);
 int mk_socket_sendv(int socket_fd, struct mk_iov *mk_io);
 int mk_socket_send(int socket_fd, const void *buf, size_t count);
 int mk_socket_read(int socket_fd, void *buf, int count);
-int mk_socket_send_file(int socket_fd, int file_fd, off_t *file_offset,
-                        size_t file_count);
+ssize_t mk_socket_send_file(int socket_fd, int file_fd, off_t *file_offset,
+                            size_t file_count);
 int mk_socket_ip_str(int socket_fd, char **buf, int size, unsigned long *len);
 int mk_socket_tcp_autocorking();
